requirements-mac-friendly.txt
# Essential packages for Mac M1/M2
langchain==0.1.0
openai==1.6.0
streamlit==1.29.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
loguru==0.7.2

# Alternative vector store (instead of faiss)
chromadb==0.4.22

# Additional useful packages
requests==2.31.0
pydantic==2.5.2
typing-extensions==4.9.0

# For embeddings
sentence-transformers==2.2.2